perf: optimize interceptor io flow

+++ ubsio-boostio/build.sh
@@ -7,7 +7,7 @@
 
 set -e
 usage() {
-    echo "Usage: $0 [ -h | -help ] [ -t | -type <build_type> ] [--ut=UT] [--cli=Diagnose] [--tp=tracepoint] [--pms=prometheus]"
+    echo "Usage: $0 [ -h | -help ] [ -t | -type <build_type> ] [--ut=UT] [--cli=Diagnose] [--tp=tracepoint] [--pms=prometheus] [--zk]"
     echo "build_type: [debug, release, clean]"
     echo "Examples:"
     echo " 1 ./build.sh -t release // 禁止添加tp功能, 对外发布包禁止添加cli功能"
@@ -24,6 +24,7 @@ BUILD_UT=OFF
 TP_FLAG=OFF
 CLI_FLAG=OFF
 PROMETHEUS_FLAG=OFF
+ZK_FLAG=OFF
 BUILD_TYPE=debug
 arch=$(uname -m)
 if [ ! -d "${BUILD_DIR}" ]; then
@@ -60,6 +61,9 @@ while true; do
             shift ;;
         --pms )
             PROMETHEUS_FLAG=ON
+            shift ;;
+        --zk )
+            ZK_FLAG=ON
             shift ;;
 		    -h | -help )
             usage
@@ -124,6 +128,12 @@ else
     CMAKE_FLAGS+="-DOPEN_PROMETHEUS=OFF "
 fi
 
+if [[ "$ZK_FLAG" == "ON" ]]; then
+    CMAKE_FLAGS+="-DOPEN_ZK=ON "
+else
+    CMAKE_FLAGS+="-DOPEN_ZK=OFF "
+fi
+
 CPU_PROCESSOR_NUM=$(($(grep processor /proc/cpuinfo | wc -l) -2))
 CMAKE_CMD="cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE $CMAKE_FLAGS $PROJ_DIR"
 BUILD_CMD="make install -j 16" # ci编译多核有问题
@@ -139,7 +149,7 @@ $BUILD_CMD || {
 	  exit 1
 }
 cd ${PROJ_DIR}/dist
-if compgen -G "3rdparty/zookeeper/lib/libzookeeper_mt.so*" > /dev/null; then
+if [[ "$ZK_FLAG" == "ON" ]] && compgen -G "3rdparty/zookeeper/lib/libzookeeper_mt.so*" > /dev/null; then
     \cp -d 3rdparty/zookeeper/lib/libzookeeper_mt.so* bio/lib/.
 fi
 
@@ -171,4 +181,3 @@ fi
 rm -rf boostio
 mv bio boostio
 tar -czvf BoostIO_1.0.0_$(uname -s)-$(arch)_${BUILD_TYPE}.tar.gz boostio
-
